Use strtod in CFF::dict_opset_t::parse_bcd
