+++ src/hb-cff-interp-dict-common.hh
@@ -27,6 +27,7 @@
 #define HB_CFF_INTERP_DICT_COMMON_HH
 
 #include "hb-cff-interp-common.hh"
+#include "hb-float.hh"
 #include <math.h>
 #include <float.h>
 
@@ -75,37 +76,6 @@ struct top_dict_values_t : dict_values_t<OPSTR>
   unsigned int  FDArrayOffset;
 };
 
-/* Compile time calculating 10^n for n = 2^i */
-constexpr double
-pow10_of_2i (unsigned int n)
-{
-  return n == 1 ? 10. : pow10_of_2i (n >> 1) * pow10_of_2i (n >> 1);
-}
-
-static const double powers_of_10[] =
-{
-  pow10_of_2i (0x100),
-  pow10_of_2i (0x80),
-  pow10_of_2i (0x40),
-  pow10_of_2i (0x20),
-  pow10_of_2i (0x10),
-  pow10_of_2i (0x8),
-  pow10_of_2i (0x4),
-  pow10_of_2i (0x2),
-  pow10_of_2i (0x1),
-};
-
-/* Works for x < 512 */
-inline double
-_hb_pow10 (unsigned int x)
-{
-  unsigned int mask = 0x100; /* Should be same with the first element  */
-  unsigned long result = 1;
-  const double *power = powers_of_10;
-  for (; mask; ++power, mask >>= 1) if (mask & x) result *= *power;
-  return result;
-}
-
 struct dict_opset_t : opset_t<number_t>
 {
   static void process_op (op_code_t op, interp_env_t<number_t>& env)
@@ -125,130 +95,67 @@ struct dict_opset_t : opset_t<number_t>
     }
   }
 
+  /* Turns CFF's BCD format into strtod understandable string */
   static double parse_bcd (byte_str_ref_t& str_ref)
   {
-    bool    neg = false;
-    double  int_part = 0;
-    uint64_t frac_part = 0;
-    uint32_t  frac_count = 0;
-    bool    exp_neg = false;
-    uint32_t  exp_part = 0;
-    bool    exp_overflow = false;
-    enum Part { INT_PART=0, FRAC_PART, EXP_PART } part = INT_PART;
+    hb_vector_t<char> buf;
     enum Nibble { DECIMAL=10, EXP_POS, EXP_NEG, RESERVED, NEG, END };
-    const uint64_t MAX_FRACT = 0xFFFFFFFFFFFFFull; /* 1^52-1 */
-    const uint32_t MAX_EXP = 0x7FFu; /* 1^11-1 */
-
-    double  value = 0.0;
     unsigned char byte = 0;
-    for (uint32_t i = 0;; i++)
+
+    if (unlikely (str_ref.in_error ())) goto fail;
+    for (unsigned i = 0;; i++)
     {
-      char d;
-      if ((i & 1) == 0)
+      char nibble;
+      if (!(i & 1))
       {
-	if (!str_ref.avail ())
-	{
-	  str_ref.set_error ();
-	  return 0.0;
-	}
+	if (!str_ref.avail ()) goto fail;
+
 	byte = str_ref[0];
 	str_ref.inc ();
-	d = byte >> 4;
+	nibble = byte >> 4;
       }
       else
-	d = byte & 0x0F;
+	nibble = byte & 0x0F;
 
-      switch (d)
+      switch (nibble)
       {
-	case RESERVED:
-	  str_ref.set_error ();
-	  return value;
+      case RESERVED:
+	goto fail;
 
-	case END:
-	  value = (double) (neg ? -int_part : int_part);
-	  if (frac_count > 0)
-	  {
-	    double frac = frac_part / _hb_pow10 (frac_count);
-	    if (neg) frac = -frac;
-	    value += frac;
-	  }
-	  if (unlikely (exp_overflow))
-	  {
-	    if (value == 0.0)
-	      return value;
-	    if (exp_neg)
-	      return neg ? -DBL_MIN : DBL_MIN;
-	    else
-	      return neg ? -DBL_MAX : DBL_MAX;
-	  }
-	  if (exp_part != 0)
-	  {
-	    if (exp_neg)
-	      value /= _hb_pow10 (exp_part);
-	    else
-	      value *= _hb_pow10 (exp_part);
-	  }
-	  return value;
+      case END:
+      {
+	const char *p = buf.arrayZ;
+	float pv;
+	if (unlikely (!parse_float (&p, p + buf.length, &pv))) goto fail;
+	return (double) pv;
+      }
 
-	case NEG:
-	  if (i != 0)
-	  {
-	    str_ref.set_error ();
-	    return 0.0;
-	  }
-	  neg = true;
-	  break;
+      case NEG:
+	buf.push ('-');
+	break;
 
-	case DECIMAL:
-	  if (part != INT_PART)
-	  {
-	    str_ref.set_error ();
-	    return value;
-	  }
-	  part = FRAC_PART;
-	  break;
+      case DECIMAL:
+	buf.push ('.');
+	break;
 
-	case EXP_NEG:
-	  exp_neg = true;
-	  HB_FALLTHROUGH;
+      case EXP_POS:
+	buf.push ('e');
+	break;
 
-	case EXP_POS:
-	  if (part == EXP_PART)
-	  {
-	    str_ref.set_error ();
-	    return value;
-	  }
-	  part = EXP_PART;
-	  break;
+      case EXP_NEG:
+	buf.push ('e');
+	buf.push ('-');
+	break;
 
-	default:
-	  switch (part) {
-	    default:
-	    case INT_PART:
-	      int_part = (int_part * 10) + d;
-	      break;
-
-	    case FRAC_PART:
-	      if (likely (frac_part <= MAX_FRACT / 10))
-	      {
-		frac_part = (frac_part * 10) + (unsigned)d;
-		frac_count++;
-	      }
-	      break;
-
-	    case EXP_PART:
-	      if (likely (exp_part * 10 + d <= MAX_EXP))
-	      {
-	      	exp_part = (exp_part * 10) + d;
-	      }
-	      else
-	      	exp_overflow = true;
-	      break;
-	  }
+      default:
+	buf.push ('0' + nibble);
+	break;
       }
     }
 
-    return value;
+fail:
+    str_ref.set_error ();
+    return .0;
   }
 
   static bool is_hint_op (op_code_t op)
